k/fetch: made response placeholder iteration order customizable
In Kafka to prevent starvation when handling fetch request for
partitions which are stored in the fetch session server moves the
partition to end whenever some data were already read for that
partition. This way partition list rotates and all partitions have
chance to be read during multiple fetch requests.

In redpanda we implemented moving partition to the end of list in fetch
session but we did not implemented changing order of iteration in
response. This might lead to assertion being triggered as fetch handler
tried to set response in the placeholder that belonged to different
partition. This situation may only take place if consumer is using long
poll technique.

When consumer requests to wait for requested number of bytes
(`fetch.min.bytes` parameter) in redpanda we may issue multiple reads.
Each consecutive read should change response placeholders iteration
order if some of the partitions were moved to the end in 'fill partition
responses' phase. This will prevent assertion from being triggered as
order of partitions when iterating over fetch session and fetch response
will be the same.

fixes: #3336

Signed-off-by: Michal Maslanka <michal@vectorized.io>