storage/test: Improve random_batch_spec defaults
In particular, idempotence is assumed `produce.cc`
`produce_handler::handle` - `hdr.producer_id >= 0`

The defaults follow those in `model::record_batch_header`, but
those fields are always explicitly written.

Change the defaults to `-1`, which is less innocuous and follows
Kafka defaults more closely.

This fixes a bug (which is yet to manifest somehow) where a random
batch is considered idempotent, which fails when
`enable_idempotence = false`.

This is needed in a future commit for testing list_offsets by time.