c/persisted_stm: fix sync timeout in case of acks=1
Previously, if the user produced with acks=1, persisted_stm::sync always
timed out because we tried to wait until commit_index reached the
current dirty offset and, even though we flushed on the leader, the
followers didn't flush and commit_index didn't move.

Instead, now we sync only up to the last offset before the start of the
current term. This upholds the contract of persisted_stm::sync ("it waits
until the state machine is caught up with all the events written by the
previous leaders") and we don't have to do any additional flushes
because after election the new leader replicates the configuration batch
and always does it with acks=-1.