cluster: use clearer wording for maintenance mode update
Every node in the cluster processes members table updates. One of these
updates is recording which node(s) are in maintenance mode. The working
"changing node N to maintenance mode" suggests an action, but in fact,
the most message is merely recording this state as metadata. Changed to
'marking node N' which communicates what is happening more clearly.

Signed-off-by: Noah Watkins <noah@redpanda.com>