cluster: handle race that evades leadership blocking
A natural race exists when blocking leadership of node that occurs if
leadership blocking races with an election. One way to solve this would
be to add more states into consensus or hold the consensus lock (or
other lock) over a much wider range of code / fibers in order to provide
a synchronization point that the drain manager could use.

Alternatively, we can accept that the race exists and treat it similarly
to how we use abort sources to shutdown services. We allow the race to
exist, but check for blocking at strategic locations. These locations
are after a pre-vote: if the pre-vote was successful but leadership is
blocked then we do not proceed to the voting phase.

If leadership blocking occurs after voting starts then if the vote would
have been successful we treat it as a failed vote.

Signed-off-by: Noah Watkins <noah@redpanda.com>