rpk: add interface to decode generic http error
When we throw http exceptions in the admin server the body contains a
json encoded dictionary with message and error code. This PR provides
a type and a decoder that can be used to decode this common type of
error. Most used to extract the message to form better error printing.

Signed-off-by: Noah Watkins <noah@redpanda.com>