tests: use --read-committed in RpkConsumer
Without this, it is quite legal for a consumer
to see the partitions offset go backward when
a leadership change happens.  This generates
an error output from RPK, which RpkConsumer
doesn't handle (and probably can't handle
in a general way).

I think most tests using this client are
intending to read committed offsets anyway.