r/vote_stm: do not step down when replicating configuration failed
In redpanda leader replicates a configuration after successful vote
round. Configuration replication success isn't required for the node to
hold the leadership.

In some scenarios, when majority of followers are behind the leader and
need recovery, configuration replication may timeout. This doesn't mean
however that leader should stepdown. In order to maintain leadership it
should continue recovering followers and step down only if there are no
heartbeat responses from majority of the followers.

Signed-off-by: Michal Maslanka <michal@vectorized.io>