rpc: fix assertion after parsing failure
It is invalid to invoke future::get_exception twice. The first call
moves the state out of the future, and the second invocation will
assert (on the now moved-from state) that the future contains an
exception but it does not.

This results in the assertion failure:

  seastar/core/future.hh:545: std::exception_ptr seastar::future_state_base::get_exception() &&: Assertion `_u.st >= state::exception_min' failed.

Signed-off-by: Noah Watkins <noah@redpanda.com>
(cherry picked from commit 2704a8b68f291e7827000c6e883088ffaaa23993)