tests: context manager to add random failures
Uses FailureInjector and the admin utility to create a
context manager which can be used to introduce random
node failures and other operations including:

1. Process kills
2. Node decommissions
3. Leadership transfer

The context manager introduced here starts a thread on
entry, which sleeps for a random interval and then applies
an operation to the cluster. The thread continues this
until the context exits. Nodes affected can be limited by
passing a configuration to the context manager, along with
time limits.