test: use rpk maintenance enable wait for barrier
Prior to this change ducktape wait_until was used to create a barrier
to wait on maintenance mode and a node being drained. having this
built-in to rpk is useful for cases like manually running an upgrade
where waiting on draining to complete is interactive, or added into a
script.

Signed-off-by: Noah Watkins <noah@redpanda.com>