tests: refactor remote script execution
This is a reusable pattern: we have a small self contained
python program that we'd like to run on a remote node.  Create
a remote_scripts/ directory to hold them, and add an
inject_remote_script helper for tests that would like
to send the script to a remote node ready for execution.

(cherry picked from commit 0dac6f49c6a68cfa590419e5c9b319e24aff87aa)