cloud_storage: Add SI space leak test
The test runs FranzGoVerifier with SI enabled. It keeps retention window
and cache size small so a lot of files gets created/deleted in the cache
directory. It uses random consumer and after the consumer is done it
checks what files are still opened by the redpanda process on every
node. The test expects that the number of opened cache files will go
down to 0 eventually.