kafka: no spin on consumer group topic create retry
When creation of consumer group topic fails we want to retry. However,
if the condition that is preventing topic creation doesn't automatically
add in a natural delay, immediately continuing will cause the fiber to
spin.

Instead, if the topic doesn't exist after an attempt at creation then we
inject a short delay before trying again.

Signed-off-by: Noah Watkins <noah@redpanda.com>