tests: refine many_partitions_test
Remove the code for running on docker: make
the environmental requirements assertions, as
this should now only ever run on release
builds on dedicated nodes.

Remove the low-count @params of the test, as
we are no longer interested in the graded check
of what size works: exactly the expected maximum
should work.

Use a startup timeout proportional to the data
payload, as for really high partition counts, the
replay gets very slow.

Adjust the node count up to 6 as we are no longer
trying to scrimp on resources in the context of
a CI run.

Increase HARD_PARTITION_LIMIT to 10k as this
should work reliably on dedicate nodes with the new
startup timeout.