tests: make PreallocNodesTest allocate on demand
This enables tests to do something else with their
node allowance earlier in a test, free the nodes,
and then use those nodes as the preallocated nodes.

Also allows freeing the nodes and allocating them
again later in a test, rather than having to hold
them throughout the lifetime of the test.