tests: leadership transfer and node decommission
adds contexts for leadership transfer and node decommission to the
disruptive actions context managers.

node decommission: the forward action decommissions a random node
in cluster using the admin api. the reverse action restarts the
node and adds it back to cluster with a new node id.

leadership transfer: a test topic is supplied to this context
which has already been created. a specific partition (0) of this
topic is the target of a leadership transfer to some other node
which holds replicas of the partition. there is no reverse action
and no restore action for this operation, as it does not reduce or
increase the cluster capacity like process kill or node decommission.