coproc: Remove some topic delete tests
- These tests all attempt to remove a materialized topic while the
coprocessor is still running.

- However due to the initial design of the system, coproc will attempt
to recreate the topic and re-populate it up until the previous high
watermark. If this was not performed there would be an inconsistency
between the coprocessors defined metadata and random commands sent to the
cluster by the user.

- The tests have been beneficial in understanding that this type of
concurrent delete can occur without any crashes.

- If a user wants to truly delete a materialized topic he/she must
shutdown the coprocessor first.

- Fixes: #3384