tests/rpk: update group describe regex
Sometimes, when kafka is the source cluster, group describe can return
a negative lag. See #4589 for more details. Until fixed, the regex
should account for a negative lag.

Also, it is possible for group describe to return '-' in the field
which means that no info was fetched for that field. Modify the regex
to account for this.