rpk: avoid listing offsets for no topics, fix group seek
For rpk group seek, if we seeked in a group that had no topics, we would
list *all* topics (as is the api for list) and then commit a group to
the beginning of *all* topics.

We now only seek if there are topics to seek, otherwise we print the
standard output that no offset was changed.

We also fix this potential seek problem in rpk group describe, but
realistically all groups have topics assigned.