tests: tweak Partition.recovered() definition
This previously reuqired that all segments have
an index and a compaction index file.  Now, these
files are only written out when there is some data
for them, rather than being created empty when a
segment roll happens.

The healthy state for a system is now for all but one
of the segment files to have an index: the currently
open segment does not have index files on disk.