tests: adjusts assertion size for deleted segment
When comparing restored data from S3 versus locally generated
baseline, the size of the baseline is reduced by default segment
size, to accomodate for the fact that the deleted segment will
not be recovered.

This change caches the size of the deleted segment on S3/minio,
and uses that size to reduce the size of baseline data, so that
if the deleted segment was smaller than the default segment size,
an assertion failure is not seen later because of that.

(cherry picked from commit 78958376f91ad7a4859c1b83e4fb5e9fcbe70931)