rm_stm: linearize fast retries
When a client retries a request immediately and doesn't give time for
the first request to be replicated it results in UNKNOWN_SERVER_ERROR.
When the first request is still inflight and its status is unknown (it
may equally succeed or fail) so Redpanda can't decide whether it should
reject or accept the follow up attempt.

Fixing the problem by introducing a per request id mutex to make the
follow up request wait until the first request is resolved.

Fixes https://github.com/redpanda-data/redpanda/issues/4366