cloud_storage: treat S3 InternalError as a backoff
S3 doesn't tell us why it's having an InternalError, but
it doesn't indicate an error on our part, so we should
treat it as retryable, _but_ with a backoff to hopefully
sidestep whatever issue the service is happening.  So
treat it like a slowdown.

(cherry picked from commit 43837704a62db0120f90a9d9b33407a4e26ba0be)