cluster: wait for health refresh if leader
Previously the health monitor would not refresh health data if it were
the leader, except for the asynchronous timer that fires periodically to
do this refresh.

The problem that we were able to identify is that if a user queries for
health information and the node is a leader, then it is ignored if the
health information needs a refresh or not. this may not seem like a big
deal (its slightly stale health information), and it is isn't, but there
is also a window of vulnerability shortly after bootup when no health
information is present. and this caused problems for rpk in some
ducktape tests that would query under the conditions that bootup had
just occured _and_ leadership had transferred to the node, thus the
query would not update the metadata on demand and return esoteric
errors about health information for nodes not existing.

Signed-off-by: Noah Watkins <noah@redpanda.com>