s3: Improve error handling
Log unexpected error replies. If the reply can't be parsed (not a valid
XML) log it limiting the size (up to 256 characters) and rate limit it
to one message per second per shard. Do this on WARN level.

Log http header for every unsuccessful reply on WARN level. Reply header
doesn't have any sensitive information so it's safe to log it.