application: fail on unexpected positional args
When the application is launched it may be supplied unexpected
positional args. Ideally this should be caught by seastar's cli
parsing tooling but because of a possible RTTI mismatch the exception
is missed and results in program termination.

With this change all positional args are collected in one place and
during validation we fail if these args are seen, this way we avoid
termination during seastar's cli parsing.