cloud_storage: add check before upload segment
when uploading segments to cloud storage, there are a couple of
points where we have to wait and we may lose leadership. the changes
introduced here add a check after these operations to cancel upload
if leadership was lost.

additionally a read lock is added to the segment being uploaded, so
that if the leadership is lost and gc kicks in, the segment is not
removed mid upload, causing EBADF.