model: move random batch generator to model module
The random model generator depends only on the model and random
libraries. However, prior to this commit the generator was kept in the
storage module. This made dependencies awkward to resolve because
storage also depends on model module.

This commit simplifies the situation by moving the random batch
generator to the model module with a single dependency on the primitive
random library.

Signed-off-by: Noah Watkins <noah@redpanda.com>