expiring_promise: maintain exception type
This is an example of an exceptional future being unwrapped and logged.
It is shown as a bare std::exception, but we should not have a single
instance in the tree of throwing a bare exception.

   WARN 2022-05-24 11:09:36,133 [shard 0] cluster - config_manager.cc:134 -
   Exception during bootstrap: std::exception (std::exception)

In this case the future's exception was set by abort source via the
expiring promise utility. However, the utility's set_exception method is
overloaded. One of the overloads, `set_exception(const std::exception&)`,
appears to be the culprit, by effectively casting to the parent type
and throwing away valuable type information by the type it becomes a
std::exception_ptr.

This commit passes throw all of the type and value classes to the
underlying seastar promise and let's it do its own thing.

Signed-off-by: Noah Watkins <noah@redpanda.com>