tests: enable round robin distribudion of records
when using kafka-producer-perf-test.sh via the KafkaCliTools utility,
provide an option to force round robin distribution of records. In
modern versions of kafka the producer by default uses sticky assignment
of records where the producer sticks to a partition until batch size is
reached or linger.ms is up. The next partition is sent messages when the
batch is full. This provides better performance in real world uses but
in some test cases we rely on messages being distributed equally among
partitions for testing and assertions, eg when triggering shadow
indexing uploads.

The option round_robin_partition enables to use the older partitioner,
which is potentially slower but ensures that each partition gets a
roughly equal amount of messages.