tests: use rpk producer for even record distribution
kafka tools cli used in some of the tests could produce non uniform
message distribution, causing errors in assertion because the segment
is not uploaded to SI.

using rpk producer produces better distribution as it uses random
keys.

this commit also fixes a log message and adds a topic specific pattern
to the matching condition of topic_manifest.json. This is to avoid the
verification matching files for topics which are not part of the test.