k8s: Change Redpanda ID calculation
For Redpanda pods that are newly created Redpanda ID will be randomly generated
For Redpanda pods that has data in persistent volume the Redpanda ID will
become statefulset ordinal number to not introducing breaking change.
For Redpanda pods that have .redpanda_id file in the Redpanda data persistent
volume the content of that file will be used for node.ID configuration.