application: handle unwanted positional args
adds an extra parse to be able to validate unwanted positional args
before the seastar CLI parsing begins.

There are a few other ways to avoid unwanted positional args but they
all suffer from drawbacks:

* Let seastar CLI parsing handle the args: this should work but due to
a type mismatch the catch clause in seastar does not recognize the
exception thrown by boost, causing program termination.

* Add a catch-all positional argument field to the CLI and then fail
later if the field is present in the parsed output: this adds an extra
help option which looks confusing.

* Explicitly instantiate the exception class in application.cc: this
causes the throw and catch to match and everything works correctly, but
the code looks confusing and is hard to maintain.

Double parsing is a little expensive at startup but allows to exit
early, before entering the seastar run function. It also allows us to
control the error messages emitted on bad input.