kafka/create_topics: validate config literals
Some of the topic creation configurations take string literals as
values. Prior to this patch creating a topic with an unexpected string
literal value failed without returning any error message due to an
unhandled exception thrown in the parsing of the literal.

This patch introduces validators which ensure that the string literal
is a valid configuration value and return a meaningful error otherwise.

Fixes #3081