kafka/p/produce: Block writes when a disk is full.
Bail out early in produce handler if we are out of disk space.

We return Broker Not Available error, since it is retryable. Note: This
behavior may change in the future to a non-retryable Storage Error.

This is the v1 implementation of full disk avoidance, which aims to
reserve enough space to keep the cluster operational by blocking produce
requests as soon as any drive in the cluster falls below the new
min_free_bytes threshold.