rpc: use variable client transport version
Adds a version field to rpc client transport and uses this version when
sending a message to the server. The new version tracked is initialized
to v0 and never changed in this patch, preserving the existing behavior.

The client transport version will be used to dynamically upgrade a
client from adl to serde encoding through client/server negotiation.

Signed-off-by: Noah Watkins <noah@redpanda.com>