rpc: use max value for unsupported version
Using a maximum value rather than something like max_supported + 1 has
the same semantic meaning when used. However the problem with
max_supported + 1 is that it makes it harder to create a patch series
that incrementally adds complexity. The reason is that it is convenient
to add the names of the new versions before the patch series fully
supports the new version. The result is an intermediate state in the
patch series in which max_supported + 1 has the same value as another
name, causing issues for existing switch statements as an example
complication.

Signed-off-by: Noah Watkins <noah@redpanda.com>