rpc: client transport handles version upgrade
Changes the client to upgrade its version when it finds that it is
connected to a peer that supports serde. Once the transport is upgraded
there may be a small window of time when the the transport version has
changed but inflight requests with a lower version are inflight.
Therefore it is important that message processing samples the version
and uses a consistent version throughout the lifetime of the messages,
rather than examining the active transport version on demand.

Signed-off-by: Noah Watkins <noah@redpanda.com>