security: don't log SASL/SCRAM bytes
While not directly sensitive, the bytes could be useful for a shrewd
attacker. It's not immediately clear when the bytes would be useful to
typical users, so this patch removes them outright.

Note that a redacted operator<< is already defined for scram_credential&
so I've removed the more detailed operator entirely.