kafka: redact operator<<s for some messages
We currently generate operator<< functions that dump the full contents
of our generated structs. This is not a good practice, as it makes it
very easy to, for instance, unintentionally log the contents of
potentially sensitive messages.

This patch takes a swing at addressing this, by labeling certain
fields (namely, SASL bytes) as sensitive, and generating a redacted
operator<< for them.

I left a TODO describing a more robust approach we might consider. For
now, this at least gives us a bit more confidence that these messages
don't leak anything sensitive when logged.

h/t Noah Watkins for the field redaction suggestions.