k8s: Extend role policy rule
As cluster is bootstrapping the node with pod ordinal 0 might lose its data
folder, the seed servers list should not be empty if cluster already was
formed. There might be a race condition where pod with ordinal 0 might form
one node cluster and as it becomes a leader it might overwrite other nodes
ring0 raft group history. To prevent such case the seed server list will
be cleared only if no other node in the cluster is working (pod in running
phase) and reports as live (v1/brokers list where is_alive is set to true).