storage: don't log errors during recovery
When parsing, an unexpected truncation maybe:
- totally normal, during recovery after an unclean
  shutdown.
- hugely worrying, under any other circumstances.

Express this distinction with a boolean flag to
continuous_batch_parser, which log_replayer may
set to true to handle the recovery case.

This improves observability in the field for
anyone searching on ERROR, and reduces the number
of cases we must allowlist in our own automated
tests.