Make partition allocator tests more generic
In one case, we relied on the exact result of the allocator in breaking
ties when nodes were tied in score: since we want to have the freedom
to choose the strategy differently (e.g., random allocation) in that
case we change the test to not assume a particular tie breaking
strategy.

Issue #4809.