cloud_storage: Add access_time_tracker
The component is using hashing to track file access times.
It's based on btree. Instead of storing file names as keys it stores
hashes. This introduces some room for conflicts but this isn't a problem
for the cache (we will overestimate access time at worst case).

The component can be serialized/deserialized using Serde.