admin_server: redact configs in admin API
Previously the admin API would send over unredacted configs.  While the
endpoint is only authorized to superusers, this is not a good practice.

There is an an important side effect of this redaction: rpk no longer
knows about the existing values of sensitive properties. This means that
it can no longer, e.g. tell the difference between an old secret
property and a new one. Importing same exact configs that already exist
will now no longer yield a no-op, and will instead result a new config
version that is identical to the old version.