cloud_storage: Maintain access times in cache
Use a dedicated data structure to maintain access times for all accesed
files in cache. The data structure is serialized periodically. The
snapshot is stored in cache directory itself. All shards are
communicating access time information to shard 0. Shard 0 maintains the
data structure and hydrates it on startup.

The access times are not stored precisely for every file. Instead the
mapping is done by the hash of the file name. This is needed to save
memory.