config: introduce redact_secrets tag
This introduces a flag that, when set to true, will redact secret
fields. It is used as an argument for property::to_json(), and by
extension, to_yaml(). The flag currently has no functionality, and is
set to not redact anywhere.

I don't expect any of our current usages to to_yaml() to require
redaction, and to that point, I could have given it a default argument.
I opted not to so developers are forced to acknowledge the redaction
decision upon using it.

(cherry picked from commit a78758fd310a2806a9d2ae76e2ff50b5c84ff561)

Conflicts:
- src/v/cluster/config_manager.cc
- src/v/config/base_property.h
- src/v/config/config_store.h
- src/v/config/property.h
- src/v/config/tests/config_store_test.cc
- src/v/redpanda/admin_server.cc
- src/v/redpanda/tests/fixture.h
- Conflicts caused rapidjson namespace being used in this branch instead
  of json. No logical changes.