rpk: remove viper and mapstructure
This commit removes viper and mapstructure
These are all the required changes to handle the
parsing, weak typing and writes to config file.

From now on we will print warning to users so they
know what needs to be changed in the config file
before we deprecate this weak typing.

This also include Yaml V3 upgrade, one notable
change is the indent in yaml files.