k/replicated_partition: do not validate fetch offset when stopping
The issues was caused by out of range error returned to the consumer
when redpanda was shutting down - to simulate cluster update. When
validating consumer offset is reset it asserts:
```
AssertionError: Consumed from an unexpected offset
```
this assertion stops consumer thread and it is not longer able to
consume messages. Tests results with failure as consumer is not able to
consume requested number of messages. This may happen in real life
scenarios when Redpanda is being shutdown gracefully. We need to prevent
shutdown sequence from resetting consumer offsets.

When partition is shutting down we may not be able to correctly
validate consumer requested offset. We claim that the offset is
valid not return `OUT_OF_RANGE_ERROR` and reset consumer offsets.

We instead return NOT_LEADER_FOR_PARTITION error to force client retry.

Fixes: #4976

Signed-off-by: Michal Maslanka <michal@vectorized.io>