tests: Deflake test_migrating_consume_offsets
Fixes synchronization around failure injection and
node restarts.

The test is structured as follows.

1. start a cluster without consumer_offsets support
2. Run an async failure injection loop that can potentially do a SIGKILL
on a random redpanda process.
3. Do a consumer groups test on a sample topic.
4. Restart the cluster to enable the new __consumer_offsets topic
feature.
5. Verify that __consumer_offsets functionality works as expected.

It turns out that (2) and (4) are conflicting with each other due to
poor synchronization. Due to some racy code these operations are
interleaved resulting in failure injector attempting to kill a
non-existent PID.

This patch makes sure that only one of these two operations run at any
given time. A given node is marked 'busy' in a thread-safe way if either
of the operations is running and the other thread has to wait.