tests: update internal topic replication factor
during disruptive test, the cluster does not appear to be able to
elect a new leader when the current leader is killed. looking at
controller logs it seems to be stuck on the old leader for the
duration of the test.

updating the internal topic replication factor to be in sync with the
number of brokers. this factor is three by default which is the same
as number of brokers in the test, but here we set it explicitly.