cloud_storage: launch auth refresh loop in remote
remote sharded service on startup launches a background fiber which
updates temporary credentials and passes them to the S3 client pool.

in case of static credentials, a wrapper is used for the static
credentials to pass to the apply credentials object but no refresh
operation is started.

existing remote tests are adjusted for the changes.