tests: fix rare failure in test_old_node_join
If this test ran after a test that did not clear
out the node's storage properly, then it could
fail to start with "configuration invariants changed".

Fix it by explicitly cleaning the node at test start:
usually we do this in RedpandaService.start(), but this
test doesn't start that one node as part of its main
call to start().