consensus: rename is_leader -> is_elected_leader
In Raft protocol when a node wins an election and becomes a leader it
doesn't mean that the data stored on the leader are final. A previous
leader could crash and the records with older term may be unconfirmed
and under-replicated.

Quotes from the Diego Ongaro's thesis:

    However, a leader cannot immediately conclude that an entry from a
    previous term is committed

    Once an entry from the current term has been committed in this way,
    then all prior entries are committed indirectly because of the Log
    Matching Property.

In Redpanda when a node becomes a leader we replicate the configuration
batch to make sure that all previous records are up replicated.

Unfortunally this subtle nuance of the consensus::is_leader (a node is
a leader but the hosted data aren't final) is confusing and already led
to an error causing a consistency violation.

We use raft group's leadership and the partition leadership backed by
that group interchangeable and when a client's request comes after a
node becomes a leader but before the replication of the configuration
is over it causes problems: under chaos tests we observed that records
from an aborted transaction became visible.

It happens because all persisted_stm based state machines (idempotency,
transacitons, shadow indexing) assume that once a raft group is a leader
its configuration batch is replicated to guarantee that records
replicated in the previous term are applied (see persisted_stm::sync).

Renaming consensus::is_leader into consensus::is_elected_leader and in-
troducing "true" consensus::is_leader to highlight the nuanced nature
of raft leadership and avoid this problem from now on.

(cherry picked from commit 10b6c48c1a3b1392175fd496363aa8a086997ef4)

    Conflicts:
        src/v/archival/ntp_archiver_service.cc
        src/v/archival/service.cc
        src/v/cluster/controller.h
        src/v/cluster/drain_manager.cc
        src/v/cluster/health_monitor_backend.cc
        src/v/cluster/metrics_reporter.cc
        src/v/cluster/partition.h
        src/v/cluster/scheduling/leader_balancer.cc
        src/v/cluster/tm_stm.cc
        src/v/coproc/partition.h
        src/v/coproc/script_context_backend.cc
        src/v/kafka/server/group_metadata_migration.cc
        src/v/kafka/server/handlers/fetch.cc
        src/v/kafka/server/partition_proxy.h
        src/v/raft/consensus.cc
        src/v/raft/consensus.h
        src/v/raft/recovery_stm.cc
        src/v/raft/replicate_batcher.cc