Extend lifetime of group for background tasks
In general the lifetime of the kafka::group objects are a bit
fraught: groups may be destroyed at any time, e.g., because
leadership of the associated group partition was moved to
another node. This means that operations dispatched on the
group object must be careful to extend the life of the group
for the duratoin of the operation.

Much of this is handled in the group_router, which usually
attaches a final continuation to futures returned by group
method which holds on to a group_ptr: this ensures that even
if the group is erased() from the main group map, it will live on
until the futures resolve.

However, in a couple of places we launch background tasks
during a group operation. These tasks don't hold to any group
pointer, so they may race with group destruction. Losing the
race manifests as segfaults or other weird behavior.

We fix this by enabling group for shared_from_this and using
that when we launch background tasks, attaching a future in
the "usual" way.

Fixes #5118.