redpanda: add a comment about arg validation
I was tempted to reorder our arg validation, placing validation outside
the run() call. It was pointed out that this has undesirable side
effects because the run() call processes args like --help that result in
valid omissions of args. This adds a comment so it's less tempting in
the future.