r/consensus: allow aborting configuration when no leader is elected
Aborting configuration change is an operation that may lead to data
loss. It must be possible to abort configuration change even if there is no
leader elected. We simply append new configuration to each of the
replicas log. If new leader will be elected using new configuration it
will eventually propagate valid configuration to all the followers.

Signed-off-by: Michal Maslanka <michal@vectorized.io>