c/topic_table: store information about previous assignment
In order to be able to abort current reconfiguration we must know what
is the previous desired state. We will keep that information in
`_update_in_progress` map. Previously it only contained the `ntp`'s that
are being moved but now it will also contain information about previous
replica set.

Signed-off-by: Michal Maslanka <michal@vectorized.io>