operator: add scale handler to properly decommission and recommission nodes
This adds a handler that correctly manages upscaling and downscaling the cluster, decommissioning nodes wheh needed.

The handler uses `status.currentReplicas` to signal the amount of replicas that all subcontrollers should materialize.
When a cluster is downscaled, the handler first tries to decommission the last node via admin API, then decreases the value of `status.currentReplicas`, to remove the node only when the cluster allows it.

In case the cluster refuses to decommission a node (e.g. min replicas on a topic higher than the desired number of nodes), the user can increase `spec.replicas` to trigger a recommission of the node.