(split) operator: always read config status from the leader to avoid inconsistent results
If we get the configuration status from any node, we may get inconsistent data since changes to the configuration take time to be propagated to all nodes. Asking the information to the leader allows us to track the propagation of changes and wait for completion correctly.

Changes that require a restart of the cluster should not be missed with this approach.