redpanda: add a test for --version
This commit adds a ducktape test to ensure the returned version from the
new --version flag is a sensible string.

The test still uses the cluster decorator to reserve the correct number
of nodes. However, the decorator previously required that the Redpanda
service be started, indicated by the presence of logs. To circumvent
this requirement, I added an argument to the decorator to allow ignoring
logs.

(cherry picked from commit 5c342cf53f372fc20b0641401550869b82ca9f94)