perf/tests: Define an OMB perf run template
Adds a ducktape perf run template that runs a standard
OMB(OpenMessagingBenchmark) workload.

This patch restructures the ducktape OMB code by
templatizing the input workload and driver files.
This makes it easy to add any new drivers/workloads
and run them as a part of the test.

The end goal is to define a realistic benchmark that
can catch perf regressions by fine tuning the workloads
and this patch lays the foundation for that.