cloud_storage: Adjust access time tracker on cleanup
This commit adds ability to adjust access_time_tracker in cache on every
cleanup. The problem is that the user can remove files from cache
manually. In this case the data will be kept in aceess_time_tracker
indefinitely. The tracker state will be saved on disk and restored on
next startup. This will lead to infinite growth of the access time
snapshot file.

To avoid this this commit adds 'remove_others' method to the
access_time_tracker. This method accepts another access_time_tracker and
removes all keys which are not present in that new tracker.

The initial cleanup procedure computes precise size of the cache
directory and cleans up the access time tracker (which is loaded from
disk at this point and may have some stale data).

The similar thing is done during cache eviction. As a side effect the
code computes the precise size of the cache dir and removes stale
entries from the access time tracker.